Some refactoring and new components
