+++ pages/_document.tsx
@@ -4,21 +4,7 @@ import { Html, Head, Main, NextScript } from "next/document";
 export default function Document() {
 	return (
 		<Html lang="sv">
-			<Head>
-				<link rel="shortcut icon" href="/images/favicon-01.png" />
-				<link
-					rel="icon"
-					type="image/png"
-					sizes="32x32"
-					href="/images/favicon-01.png"
-				/>
-				<link
-					rel="icon"
-					type="image/png"
-					sizes="16x16"
-					href="/images/favicon-02.png"
-				/>
-			</Head>
+			<Head />
 			<body>
 				<Main />
 				<NextScript />
